started refactoring to typescript and edit front page
