+++ game-editor/frontend/index.html
@@ -7,7 +7,7 @@
 </head>
 <body>
 <div id="root"></div>
-<script src="./src/main.jsx" type="module"></script>
+<script src="./src/main.tsx" type="module"></script>
 </body>
 </html>
