+++ game-editor/frontend/package.json.md5
@@ -1,1 +1,1 @@
-7ac6284c0663259ab8187b439b6a0808
+5e0e99e4994909f3bb05d1141c7cc179
