+++ game-editor/frontend/src/style.css
@@ -3,7 +3,7 @@
 @tailwind utilities;
 
 html {
-	background-color: rgba(27, 38, 54, 1);
+	background-color: rgba(72, 77, 132, 1);
 	text-align: center;
 	color: white;
 }
