+++ game-editor/frontend/tailwind.config.js
@@ -1,10 +1,9 @@
 /** @type {import('tailwindcss').Config} */
+
+const colors = require("tailwindcss/colors");
+
 export default {
-	content: [
-		"./index.html",
-		"./src/**/*.{js,ts,jsx,tsx}",
-		"./src/*.{js,ts,jsx,tsx}",
-	],
+	content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
 	theme: {
 		screens: {
 			sm: "480px",
@@ -12,7 +11,12 @@ export default {
 			lg: "976px",
 			xl: "1440px",
 		},
-		extend: {},
+		extend: {
+			colors: {
+				blueWhale: "#182A3A",
+				wildBlueYonder: "#7A89C2",
+			},
+		},
 	},
 	plugins: [],
 };
